refactor(router): move matchPattern into router
