+++ packages/router/src/index.ts
@@ -1,6 +1,5 @@
-export { matchPattern } from './matcher';
-export type { Match } from './matcher';
-export { Router } from './router';
+export { Router, matchPattern } from './router';
+export type { Match } from './router';
 export { Route } from './route';
 export { Link } from './link';
 export { Redirect } from './redirect';
