--- packages/router/src/matcher.ts
@@ -1,26 +1,0 @@
-export interface Match {
-  params: Record<string, string>;
-}
-
-export function matchPattern(pattern: string, path: string): Match | null {
-  const patternParts = split(pattern);
-  const pathParts = split(path);
-
-  if (patternParts.length !== pathParts.length) return null;
-
-  const params: Record<string, string> = {};
-
-  for (let i = 0; i < patternParts.length; i++) {
-    const p = patternParts[i];
-    const v = pathParts[i];
-    if (p.startsWith(':')) params[p.slice(1)] = v;
-    else if (p.toLowerCase() !== v.toLowerCase()) return null;
-  }
-
-  return { params };
-}
-
-function split(path: string) {
-  const trimmed = path.replace(/^\/+|\/+$/g, '');
-  return trimmed === '' ? [] : trimmed.split('/');
-}
