+++ packages/router/src/router.ts
@@ -1,8 +1,11 @@
 import { Component } from '@expressive/react';
 import { Children, ReactElement, ReactNode, isValidElement } from 'react';
 
-import { matchPattern } from './matcher';
 import { Route } from './route';
+
+export interface Match {
+  params: Record<string, string>;
+}
 
 export class Router extends Component {
   path = window.location.pathname;
@@ -30,6 +33,24 @@ export class Router extends Component {
   }
 }
 
+export function matchPattern(pattern: string, path: string): Match | null {
+  const patternParts = split(pattern);
+  const pathParts = split(path);
+
+  if (patternParts.length !== pathParts.length) return null;
+
+  const params: Record<string, string> = {};
+
+  for (let i = 0; i < patternParts.length; i++) {
+    const p = patternParts[i];
+    const v = pathParts[i];
+    if (p.startsWith(':')) params[p.slice(1)] = v;
+    else if (p.toLowerCase() !== v.toLowerCase()) return null;
+  }
+
+  return { params };
+}
+
 export function resolve(children: ReactNode, path: string): ReactNode {
   for (const child of Children.toArray(children)) {
     if (!isRouteElement(child)) continue;
@@ -40,4 +61,9 @@ export function resolve(children: ReactNode, path: string): ReactNode {
 
 function isRouteElement(node: unknown): node is ReactElement<{ to?: string }> {
   return isValidElement(node) && node.type === Route;
+}
+
+function split(path: string) {
+  const trimmed = path.replace(/^\/+|\/+$/g, '');
+  return trimmed === '' ? [] : trimmed.split('/');
 }
